test auth and firestore 100%
