+++ src/controller/controller-auth.js
@@ -10,16 +10,6 @@ export const signInForGoogle = () => {
   const provider = new firebase.auth.GoogleAuthProvider();
   return auth.signInWithPopup(provider);
 };
-// -------------------------------------CHECK SESSION ACTIVE----------------------------------
-export const checkSesionActive = ChangeView => firebase.auth().onAuthStateChanged((user) => {
-  let route = '';
-  if (window.location.hash === '#/signUp') route = '#/signUp';
-  if (window.location.hash === '#/recoverPassword') route = '#/recoverPassword';
-  if (user) {
-    route = window.location.hash;
-  }
-  return ChangeView(route);
-});
 // ---------------------------------------------SIGN UP---------------------------------------
 // Create user
 export const createUser = (email, password) => {
